refs #74 Separate filter definitions
